Forecast

+++ app/src/main/java/com/example/android/weatherapp/model/Main.java
@@ -12,7 +12,7 @@ public class Main {
 
     @SerializedName("pressure")
     @Expose
-    private int pressure;
+    private float pressure;
 
     @SerializedName("humidity")
     @Expose
@@ -34,11 +34,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
